FiatApi: fetch CoinGecko requests sequentially instead of in parallel
This avoids having many parallel CoinGecko requests waiting and retrying continuously
and in parallel on the low rate limit. While sending requests while the rate limit is
hit, does not negatively affect the rate limit, it is eventually just a lot of unne-
cesssary requests and noise. Note that sending the requests sequentially does not waste
any of our limit, as the bottleneck is the rate limit and not the sequential execution
of requests.