FiatApi: shorten names of currency related enums, types and constants
The names have become a bit too long over time, as new variations like bridged currencies
were introduced, and would become even longer with the upcoming differentiation of different
api providers, like CryptoCompare and CoinGecko. As this change will lead to breaking changes
in the API anyways, now is a good time to also change the naming scheme for the enums, types
and constants, which is also a breaking change.