FiatApi: add all bridgeable currencies, incl. overlap with provider support
Change BridgedFiatCurrency to BridgeableFiatCurrency enum, which is not only a name change but
also a change in semantics. BridgeableFiatCurrency now includes all currencies bridgeable by any
of our bridges, also including currencies which are already supported by a provider directly
without the need to use a bridge. Previously, currencies in BridgedFiatCurrency and FiatCurrency
supported by Coingecko as single provider were mutually exclusive, i.e. there was no overlap.
This change is now being introduced in light of CryptoCompare being added as a second provider,
with a different set of supported currencies, due to which the mutual exclusivity is not possible
anymore. With the introduction of a second provider, whether a currency is considered a bridged
currency (for that provider) will depend on the provider. Instead, two type guards are provided
now: isBridgeableFiatCurrency and isBridgedFiatCurrency, where isBridgedFiatCurrency will depend
on the provider as soon as the second provider is added.