FiatApi: limit parallel CryptoCompare requests on daily/monthly limit
We ignore daily and monthly limits in hopes of the usage being reset earlier than on
day or month reset, for example by an IP change, but now limit parallel requests when
the daily or monthly limit is hit to avoid unnecessary parallel requests. The parallel
limit is reset on the next successful request.