RateLimitScheduler: replace period reset delay by stricter period reset offset
The delaying of period reset times covered accounting for clock differences between system
time and server time for web requests to some degree, if the queue is full and worked on
over multiple periods, but it could still lead to rate limiting, for example if the queue
is empty and new tasks get scheduled within the delay period of the previous period, which
already falls into the next actual period, then waiting for the delayed period to end and
scheduling additional tasks in the new period, which is the same actual period and can thus
lead to unexpected rate limits.
Such problems are now circumvented by a stricter safety buffer before and after period reset
times in which no tasks are allowed at all, to avoid uncertainties of into which actual
period such tasks fall on the server.