Use USDT for historical NIM prices from CryptoCompare
By default, CryptoCompare calculates the NIM price in USD via BTC sats. But it doesn't use decimals for sats, so when the NIM price is 3 sats (rounded), then the historical prices follow Bitcoin's exactly. Using USDT as the target currency (vsCurrency) works because NIM has a direct pair with USDT on a major exchange, so CryptoCompare can use that directly.