FiatApi: support historic rates at a custom time resolution
For CryptoCompare and CryptoCompareLegacy support fetching historic rates at a
custom time resolution by providing a custom interval and additional aggregation
of intervals. This can be used to reduce the returned data points, or required
requests, at the expense of less detailed / precice results, or to request data
at higher resolution.

The custom parameters are available for both, getHistoricExchangeRates, where they
affect the precision of the underlying data, and getHistoricExchangeRatesByRange,
where they affect the number of fetched and returned samples.